# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flipzy API",
        "description": "Flipzy Property Management Platform API - Complete API documentation for property listings, messaging, analytics, and admin features.\n\n## Authentication\n\nThis API uses Laravel Passport for authentication. To authenticate:\n\n1. Use the `/register` or `/login` endpoint to get an access token\n2. Copy the `access_token` from the response\n3. Click the 'Authorize' button at the top of this page\n4. Enter: `Bearer YOUR_ACCESS_TOKEN` (replace YOUR_ACCESS_TOKEN with your actual token)\n5. Click 'Authorize' and then 'Close'\n\nAll protected endpoints will now use this token automatically.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/api/v1/refunds": {
            "post": {
                "tags": [
                    "Refunds"
                ],
                "summary": "Create refund",
                "description": "Create a full or partial refund for a completed transaction. Only the transaction owner or admin can create refunds.",
                "operationId": "9b7f5e7dfb31fa9198d93fbcabc75be5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "transaction_id"
                                ],
                                "properties": {
                                    "transaction_id": {
                                        "description": "Transaction UUID to refund",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "019a5882-c272-71eb-97c4-3d81aa7a4807"
                                    },
                                    "amount": {
                                        "description": "Partial refund amount (optional, defaults to full refund)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 15
                                    },
                                    "reason": {
                                        "description": "Reason for refund",
                                        "type": "string",
                                        "enum": [
                                            "duplicate",
                                            "fraudulent",
                                            "requested_by_customer"
                                        ],
                                        "example": "requested_by_customer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Refund created successfully"
                    },
                    "400": {
                        "description": "Validation error or refund not allowed"
                    },
                    "404": {
                        "description": "Transaction not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/refunds/{refund}": {
            "get": {
                "tags": [
                    "Refunds"
                ],
                "summary": "Get refund details",
                "description": "Get detailed information about a specific refund transaction.",
                "operationId": "8b9f31cc7ee097413588c13a187abdb7",
                "parameters": [
                    {
                        "name": "refund",
                        "in": "path",
                        "description": "Refund transaction UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Refund retrieved successfully"
                    },
                    "404": {
                        "description": "Refund not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your access token in the format: Bearer YOUR_ACCESS_TOKEN\n\nGet your token by:\n- Registering a new account via POST /register\n- Logging in via POST /login\n\nThe token will be returned in the response as 'access_token'.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}